+++ test/react/rails/pages_controller_test.rb
@@ -1,8 +1,8 @@
 require 'test_helper'
 
 class PagesControllerTest < ActionController::TestCase
-  test 'should not raise react_component undefined in ActionController::TestCase' do
+  test 'renders successfully' do
     get :show, id: 1
-    assert_not_raised ActionView::Template::Error
+    assert_equal(200, response.status)
   end
 end